Split resource into own files
